Refactor package structure and tests
Generate the package version from the git version.

Move source code into a src directory.

Remove redundant code from __init__ files and remove those where
possible.

Rename ml_warehouse.ml_warehouse_schema to ml_warehouse.schema to
avoid tautology.

Move fixture code to conftest.py where possible, to take advantage of
pytest auto-importing.

Move examples into the tests directory.

Remove redundant text from some test messages.

Add comments on setting up local MySQL servers for testing.

Rename test.ini to testdb.ini.
