+++ .github/workflows/ci.yml
@@ -52,8 +52,9 @@ jobs:
       - name: "Check code formatting [Black]"
         run: >-
           black --check --diff --quiet .
-          --force-exclude ml_warehouse/ml_warehouse_schema.py
+          --force-exclude src/ml_warehouse/schema.py
 
       - name: "Run unit tests"
         run: |
+          export PYTHONPATH=$PWD/src:$PWD/tests:$PYTHONPATH
           pytest --it
