--- tests/__init__.py
@@ -1,6 +1,0 @@
-import logging
-
-logging.basicConfig(
-    format="%(asctime)s %(levelname)s %(module)s - " "%(funcName)s - %(message)s",
-    level=logging.DEBUG,
-)
