+++ tests/examples/long_illumina.py
@@ -19,11 +19,13 @@
 
 from datetime import datetime, timedelta
 from typing import Sequence
-from sqlalchemy.orm import Session, Query
+
+from sqlalchemy.orm import Query, Session
 from sqlalchemy.sql.functions import func
 from sqlalchemy.sql.schema import Column
 from sqlalchemy.types import INTEGER
-from ml_warehouse.ml_warehouse_schema import (
+
+from ml_warehouse.schema import (
     IseqFlowcell,
     IseqProductMetrics,
     IseqRunStatus,
