--- tests/test.ini
@@ -1,6 +1,0 @@
-[MySQL]
-user = test
-password = test
-ip_address = 127.0.0.1
-port = 3306
-schema = mlwarehouse
