+++ tests/testdb.ini
@@ -1,0 +1,19 @@
+# Credentials to be used for unit tests.
+#
+# GitHub Actions CI should set up a MySQL service matching these credentials.
+# If you are running tests on your local development machine, you can start a
+# suitable Docker container like so:
+#
+# docker run --network host -d \
+#     -e MYSQL_RANDOM_ROOT_PASSWORD=yes \
+#     -e MYSQL_TCP_PORT=3306 \
+#     -e MYSQL_USER=test \
+#     -e MYSQL_PASSWORD=test \
+#     -e MYSQL_DATABASE=mlwarehouse "mysql:$MYSQL_VERSION"
+#
+[MySQL]
+user = test
+password = test
+ip_address = 127.0.0.1
+port = 3306
+schema = mlwarehouse
